+++ modules/tinymce/src/core/main/ts/api/Formatter.ts
@@ -41,7 +41,7 @@ interface Formatter extends FormatRegistry {
   matchAll: (names: string[], vars?: FormatVars) => string[];
   matchNode: (node: Node, name: string, vars?: FormatVars, similar?: boolean) => Format | undefined;
   canApply: (name: string) => boolean;
-  formatChanged: (names: string, callback: FormatChanged.FormatChangeCallback, similar?: boolean) => { unbind: () => void };
+  formatChanged: (names: string, callback: FormatChanged.FormatChangeCallback, similar?: boolean, value?: FormatVars) => { unbind: () => void };
   getCssText: (format: string | Format) => string;
 }
 
@@ -187,8 +187,9 @@ const Formatter = (editor: Editor): Formatter => {
      * @param {String} formats Comma separated list of formats to check for.
      * @param {function} callback Callback with state and args when the format is changed/toggled on/off.
      * @param {Boolean} similar True/false state if the match should handle similar or exact formats.
+     * @param {Object} vars Restrict the format being watched to only match if the variables applied are equal to vars.
      */
-    formatChanged: (formats: string, callback: FormatChanged.FormatChangeCallback, similar?: boolean) => Rtc.formatChanged(editor, formatChangeState, formats, callback, similar),
+    formatChanged: (formats: string, callback: FormatChanged.FormatChangeCallback, similar?: boolean, vars?: FormatVars) => Rtc.formatChanged(editor, formatChangeState, formats, callback, similar, vars),
 
     /**
      * Returns a preview css text for the specified format.